# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KETANTECH VPN Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List current user's active VPN accounts
      security:
        - cookieAuth: []
      responses:
        "200":
          description: VPN account list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VpnAccount"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get VPN account detail
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpnAccount"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /accounts/{id}/renew:
    post:
      operationId: renewAccount
      tags: [accounts]
      summary: Retired static product renewal endpoint
      deprecated: true
      description: Static product renewals are retired. Create a new account through POST /dynamic-vpn/orders.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "410":
          description: Static renewals are retired; create a new account through POST /dynamic-vpn/orders.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Dashboard summary ─────────────────────────────────────────────────
...
  /accounts/{id}/renew-dynamic/quote:
    post:
      operationId: quoteDynamicAccountRenewal
      tags: [accounts, dynamic-vpn]
      summary: Quote a balance-funded renewal for a dynamic VPN account
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/AccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DynamicRenewalBody"
      responses:
        "200":
          description: Dynamic renewal price quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DynamicRenewalQuote"
        "400":
          description: Invalid duration or account cannot be renewed dynamically
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /accounts/{id}/sync-provider:
    post:
      operationId: syncDynamicAccountProvider
      tags: [accounts, dynamic-vpn]
      summary: Synchronize a NadiaVPN dynamic account from its provider
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "200":
          description: Synchronized VPN account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpnAccount"
        "400":
          description: Account is not a synchronizable NadiaVPN dynamic account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /accounts/{id}/renew-dynamic:
    post:
      operationId: renewDynamicAccount
      tags: [accounts, dynamic-vpn]
      summary: Renew a dynamic VPN account using balance
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/AccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DynamicRenewalBody"
      responses:
        "200":
          description: Dynamic account renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DynamicRenewalResponse"
        "400":
          description: Invalid renewal request or insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Renewal already in progress for this account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Dynamic renewal failed after the request was accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Dynamic VPN ───────────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  parameters:
    AccountId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        replacement:
          type: string
...
    DynamicDurationType:
      type: string
      enum: [day, week, month]
...
    DynamicRenewalBody:
      type: object
      required: [durationType, duration]
      properties:
        durationType:
          $ref: "#/components/schemas/DynamicDurationType"
        duration:
          type: integer
          minimum: 1
    DynamicRenewalQuote:
      type: object
      required: [amount, baseAmount, resellerDiscountAmount, unitPrice, durationType, duration, durationLabel]
      properties:
        amount:
          type: number
        baseAmount:
          type: number
        resellerDiscountAmount:
          type: number
        unitPrice:
          type: number
        durationType:
          $ref: "#/components/schemas/DynamicDurationType"
        duration:
          type: integer
        durationLabel:
          type: string
    DynamicRenewalResponse:
      type: object
      required: [account, amount, discountAmount]
      properties:
        account:
          $ref: "#/components/schemas/VpnAccount"
        amount:
          type: number
        discountAmount:
          type: number
...
    VpnAccount:
      type: object
      required: [id, userId, orderId, dynamicOrder, protocol, username, serverId, server, configLink, allLinks, expiresAt, quota, usedQuota, productName, isActive, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        orderId:
          type: integer
          nullable: true
        dynamicOrder:
          type: object
          nullable: true
          required: [id, provider, providerServerId, serverDisplayName, providerAccountId, dynamicServerId, renewEnabled, supportedTypes, sellPricePerDay, sellPricePerWeek, sellPricePerMonth]
          properties:
            id:
              type: integer
            provider:
              type: string
            providerServerId:
              type: string
            serverDisplayName:
              type: string
            providerAccountId:
              type: string
              nullable: true
            dynamicServerId:
              type: integer
              nullable: true
            renewEnabled:
              type: boolean
            supportedTypes:
              type: array
              items:
                $ref: "#/components/schemas/DynamicDurationType"
            sellPricePerDay:
              type: number
            sellPricePerWeek:
              type: number
            sellPricePerMonth:
              type: number
        protocol:
          type: string
          enum: [ssh, vmess, vless, trojan, shadowsocks]
        username:
          type: string
        password:
          type: string
          nullable: true
        uuid:
          type: string
          nullable: true
        serverId:
          type: integer
        server:
          $ref: "#/components/schemas/PublicServer"
          nullable: true
        configLink:
          type: string
          nullable: true
          description: Primary config link (TLS)
        allLinks:
          type: object
          nullable: true
          description: All config link variants (tls, none, grpc, uptls, upntls)
          additionalProperties:
            type: string
            nullable: true
        expiresAt:
          type: string
          format: date-time
        quota:
          type: number
          nullable: true
        usedQuota:
          type: number
          nullable: true
        productName:
          type: string
          nullable: true
          description: Nama paket produk yang dibeli
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    PublicServer:
      type: object
      required: [id, name, location, flag, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        host:
          type: string
          nullable: true
        location:
          type: string
        flag:
          type: string
          description: Country flag emoji or code
        isActive:
          type: boolean
...